Redid control flow to work better in a many trac environment
Previously it instantiated a trac environment for every peer directory
and did the permission check before including it in the list of tracs
that a ticket can be moved to. Now it just finds everything that looks
like a peer trac directory (has a conf/trac.ini file) and then does the
permission check on the target immediately before actually moving.

Additionally we only require the "TICKET_CREATE" permission on the target trac.